% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mExplorer.R
\docType{data}
\name{small_test_dframe}
\alias{small_test_dframe}
\title{Small sample of predictor data for testing m:Explorer}
\format{A data frame with 10 observations of 18 variables}
\usage{
data(mExplorer_small_test_data)
}
\description{
Small sample of predictor data for testing m:Explorer
}
\keyword{datasets}
