% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mExplorer.R
\docType{data}
\name{small_test_response_vec}
\alias{small_test_response_vec}
\title{Small vector of yeast transcription factors for testing m:Explorer}
\format{A named character vector with 4 elements}
\usage{
data(mExplorer_small_test_data)
}
\description{
Small vector of yeast transcription factors for testing m:Explorer
}
\keyword{datasets}
