% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mExplorer.R
\docType{data}
\name{yeastCCgenes}
\alias{yeastCCgenes}
\title{Example vector of yeast transcription factors for m:Explorer}
\format{A named character vector with 186 elements}
\usage{
data(yeastCCgenes)
}
\description{
Example vector of yeast transcription factors for m:Explorer
}
\keyword{datasets}
