% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mExplorer.R
\docType{data}
\name{yeastTFdata}
\alias{yeastTFdata}
\title{Example predictor data for m:Explorer}
\format{A data frame with 6253 observations of 18 variables}
\usage{
data(yeastTFdata)
}
\description{
Example predictor data for m:Explorer
}
\keyword{datasets}
